Address BH fast tilize PR feedback
