+++ models/demos/deepseek_v3/tests/fused_op_unit_tests/moe/tilize_16x32/kernels/tilize_16x32_kernel.cpp
@@ -28,6 +28,6 @@ void kernel_main() {
     fast_tilize_init(in_cb, block_ct_dim, out_cb);
     fast_tilize_block(in_cb, block_ct_dim, out_cb, 0, 0);
     cb_pop_front(in_cb, block_ct_dim);
-    fast_tilize_uninit(in_cb, out_cb);
+    fast_tilize_uninit(in_cb, out_cb, block_ct_dim);
 #endif
 }
