+++ tt_metal/tt-llk/tests/sources/fast_tilize_metal_api_test.cpp
@@ -147,8 +147,8 @@ void run_kernel(RUNTIME_PARAMETERS params)
 
 #ifdef LLK_TRISC_PACK
 
+#include "experimental/llk_pack_fast_tilize.h"
 #include "llk_pack_common.h"
-#include "llk_pack_fast_tilize.h"
 
 void run_kernel(RUNTIME_PARAMETERS params)
 {
